split MockFileTests into some smaller TestClasses
