+++ TestHelpers.Tests/MockFileOpenTests.cs
@@ -1,0 +1,181 @@
+namespace System.IO.Abstractions.TestingHelpers.Tests
+{
+    using Collections.Generic;
+
+    using NUnit.Framework;
+
+    using XFS = MockUnixSupport;
+
+    public class MockFileOpenTests {
+        [Test]
+        public void MockFile_Open_ThrowsOnCreateNewWithExistingFile()
+        {
+            string filepath = XFS.Path(@"c:\something\already\exists.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData> 
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            Assert.Throws<IOException>(() => filesystem.File.Open(filepath, FileMode.CreateNew));
+        }
+
+        [Test]
+        public void MockFile_Open_ThrowsOnOpenWithMissingFile()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>());
+
+            Assert.Throws<FileNotFoundException>(() => filesystem.File.Open(filepath, FileMode.Open));
+        }
+
+        [Test]
+        public void MockFile_Open_ThrowsOnTruncateWithMissingFile()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>());
+
+            Assert.Throws<FileNotFoundException>(() => filesystem.File.Open(filepath, FileMode.Truncate));
+        }
+
+        [Test]
+        public void MockFile_Open_CreatesNewFileFileOnCreate()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>());
+
+            var stream = filesystem.File.Open(filepath, FileMode.Create);
+
+            Assert.That(filesystem.File.Exists(filepath), Is.True);
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void MockFile_Open_CreatesNewFileFileOnCreateNew()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>());
+
+            var stream = filesystem.File.Open(filepath, FileMode.CreateNew);
+
+            Assert.That(filesystem.File.Exists(filepath), Is.True);
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void MockFile_Open_OpensExistingFileOnAppend()
+        {
+            string filepath = XFS.Path(@"c:\something\does\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            var stream = filesystem.File.Open(filepath, FileMode.Append);
+            var file = filesystem.GetFile(filepath);
+            
+            Assert.That(stream.Position, Is.EqualTo(file.Contents.Length));
+            Assert.That(stream.Length, Is.EqualTo(file.Contents.Length));
+
+            stream.Seek(0, SeekOrigin.Begin);
+
+            byte[] data;
+            using (var br = new BinaryReader(stream))
+                data = br.ReadBytes((int)stream.Length);
+
+            CollectionAssert.AreEqual(file.Contents, data);
+        }
+
+        [Test]
+        public void MockFile_Open_OpensExistingFileOnTruncate()
+        {
+            string filepath = XFS.Path(@"c:\something\does\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            var stream = filesystem.File.Open(filepath, FileMode.Truncate);
+            var file = filesystem.GetFile(filepath);
+
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(0));
+            Assert.That(file.Contents.Length, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void MockFile_Open_OpensExistingFileOnOpen()
+        {
+            string filepath = XFS.Path(@"c:\something\does\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            var stream = filesystem.File.Open(filepath, FileMode.Open);
+            var file = filesystem.GetFile(filepath);
+
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(file.Contents.Length));
+
+            byte[] data;
+            using (var br = new BinaryReader(stream))
+                data = br.ReadBytes((int)stream.Length);
+
+            CollectionAssert.AreEqual(file.Contents, data);
+        }
+
+        [Test]
+        public void MockFile_Open_OpensExistingFileOnOpenOrCreate()
+        {
+            string filepath = XFS.Path(@"c:\something\does\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            var stream = filesystem.File.Open(filepath, FileMode.OpenOrCreate);
+            var file = filesystem.GetFile(filepath);
+
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(file.Contents.Length));
+
+            byte[] data;
+            using (var br = new BinaryReader(stream))
+                data = br.ReadBytes((int)stream.Length);
+
+            CollectionAssert.AreEqual(file.Contents, data);
+        }
+
+        [Test]
+        public void MockFile_Open_CreatesNewFileOnOpenOrCreate()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>());
+
+            var stream = filesystem.File.Open(filepath, FileMode.OpenOrCreate);
+
+            Assert.That(filesystem.File.Exists(filepath), Is.True);
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void MockFile_Open_OverwritesExistingFileOnCreate()
+        {
+            string filepath = XFS.Path(@"c:\something\doesnt\exist.txt");
+            var filesystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { filepath, new MockFileData("I'm here") }
+            });
+
+            var stream = filesystem.File.Open(filepath, FileMode.Create);
+            var file = filesystem.GetFile(filepath);
+
+            Assert.That(stream.Position, Is.EqualTo(0));
+            Assert.That(stream.Length, Is.EqualTo(0));
+            Assert.That(file.Contents.Length, Is.EqualTo(0));
+        }
+    }
+}
