+++ TestHelpers.Tests/MockFileReadAllLinesTests.cs
@@ -1,0 +1,55 @@
+namespace System.IO.Abstractions.TestingHelpers.Tests
+{
+    using Collections.Generic;
+
+    using NUnit.Framework;
+
+    using Text;
+
+    using XFS = MockUnixSupport;
+
+    public class MockFileReadAllLinesTests {
+        [Test]
+        public void MockFile_ReadAllLines_ShouldReturnOriginalTextData()
+        {
+            // Arrange
+            var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { XFS.Path(@"c:\something\demo.txt"), new MockFileData("Demo\r\ntext\ncontent\rvalue") },
+                { XFS.Path(@"c:\something\other.gif"), new MockFileData(new byte[] { 0x21, 0x58, 0x3f, 0xa9 }) }
+            });
+
+            var file = new MockFile(fileSystem);
+
+            // Act
+            var result = file.ReadAllLines(XFS.Path(@"c:\something\demo.txt"));
+
+            // Assert
+            CollectionAssert.AreEqual(
+                new[] { "Demo", "text", "content", "value" },
+                result);
+        }
+
+        [Test]
+        public void MockFile_ReadAllLines_ShouldReturnOriginalDataWithCustomEncoding()
+        {
+            // Arrange
+            string text = "Hello\r\nthere\rBob\nBob!";
+            var encodedText = Encoding.BigEndianUnicode.GetBytes(text);
+            var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { XFS.Path(@"c:\something\demo.txt"), new MockFileData(encodedText) }
+            });
+
+            var file = new MockFile(fileSystem);
+
+            // Act
+            var result = file.ReadAllLines(XFS.Path(@"c:\something\demo.txt"), Encoding.BigEndianUnicode);
+
+            // Assert
+            CollectionAssert.AreEqual(
+                new [] { "Hello", "there", "Bob", "Bob!" },
+                result);
+        }
+    }
+}
