+++ TestHelpers.Tests/TestHelpers.Tests.csproj
@@ -83,11 +83,19 @@
     <Compile Include="FileSystemTests.cs" />
     <Compile Include="MockDirectoryInfoTests.cs" />
     <Compile Include="MockDirectoryTests.cs" />
+    <Compile Include="MockFileAppendAllTextTests.cs" />
+    <Compile Include="MockFileCopyTests.cs" />
+    <Compile Include="MockFileCreateTests.cs" />
+    <Compile Include="MockFileExistsTests.cs" />
     <Compile Include="MockFileInfoFactoryTests.cs" />
     <Compile Include="MockFileInfoTests.cs" />
+    <Compile Include="MockFileMoveTests.cs" />
+    <Compile Include="MockFileOpenTests.cs" />
+    <Compile Include="MockFileReadAllLinesTests.cs" />
     <Compile Include="MockFileStreamTests.cs" />
     <Compile Include="MockFileSystemTests.cs" />
     <Compile Include="MockFileTests.cs" />
+    <Compile Include="MockFileWriteAllTextTests.cs" />
     <Compile Include="MockPathTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MockUnixSupportTests.cs" />
